refactor: Mejorar la estructura de datos y la paginación en los controladores, y actualizar el tipo de ID en Terminal

+++ src/main/java/com/cibertec/demo/modelo/Terminal.java
@@ -11,7 +11,7 @@ public class Terminal {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_terminal")
-    private Integer idTerminal;
+    private Long idTerminal;
 
     @Column(name = "nombre_ubicacion", nullable = false, length = 100)
     private String nombreUbicacion;
